requirements.txt
certifi==2019.6.16
Django==2.2.4
django-haystack==2.8.1
elasticsearch==5.5.3
pytz==2019.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
